: 398048: file modified

+++ ext/marketplace/marketplacefacades/src/com/tisl/mpl/search/converters/populator/MplSearchResultProductPopulator.java
@@ -227,7 +227,7 @@ public class MplSearchResultProductPopulator extends SearchResultVariantProductP
 			final double calculatedPerSavings = (savingsAmt / mrpPriceValue.doubleValue()) * 100;
 			//final double roundedOffValuebefore = Math.round(calculatedPerSavings * 100.0) / 100.0;
 			final double floorValue = Math.floor((calculatedPerSavings * 100.0) / 100.0);
-			final PriceData priceData = getPriceDataFactory().create(PriceDataType.BUY, BigDecimal.valueOf(floorValue),
+			final PriceData priceData = getPriceDataFactory().create(PriceDataType.BUY, BigDecimal.valueOf((int) floorValue),
 					getCommonI18NService().getCurrentCurrency());
 
 			target.setSavingsOnProduct(priceData);
@@ -254,20 +254,20 @@ public class MplSearchResultProductPopulator extends SearchResultVariantProductP
 	/*
 	 * @Override protected void addImageData(final SearchResultValueData source, final String imageFormat, final String
 	 * mediaFormatQualifier, final ImageDataType type, final List<ImageData> images) {
-	 *
+	 * 
 	 * final Object imgObj = getValue(source, "img-" + mediaFormatQualifier); List<String> imgList = new ArrayList(); if
 	 * (imgObj instanceof ArrayList) { imgList = (List) imgObj; } else { final String imgStr = (String) imgObj;
 	 * imgList.add(imgStr); }
-	 *
-	 *
+	 * 
+	 * 
 	 * if (!imgList.isEmpty()) { for (int i = 0; i < imgList.size(); i++) { final ImageData imageSearchData =
 	 * createImageData(); imageSearchData.setImageType(type); imageSearchData.setFormat(imageFormat);
 	 * imageSearchData.setUrl(imgList.get(i)); images.add(imageSearchData);
-	 *
-	 *
+	 * 
+	 * 
 	 * }
-	 *
-	 *
+	 * 
+	 * 
 	 * } }
 	 */
